requirements.txt
torch==1.5.0
torchvision==0.6.0
tensorboardX
easydict
pandas
numpy
scipy
matplotlib
Pillow
opencv-python
wget